refactor: uikit 모듈 경로 통일화
