+++ src/component/timeBlock/molecules/SelectDate.tsx
@@ -2,7 +2,7 @@ import { styled } from 'styled-components';
 import { useRecoilValue } from 'recoil';
 import { dateListState } from '@/store/atoms';
 import { DuringDate } from '@/component/timeBlock/atoms/DuringDate';
-import { BackButton, NextButton } from '@/component/uikit/atom';
+import { BackButton, NextButton } from '@/component/uikit';
 
 interface DateListProps {
   page: number;
