+++ src/component/uikit/atom/index.ts
@@ -1,14 +1,19 @@
+export * from './Back';
+export * from './BackButton';
 export * from './ButtonLarge';
 export * from './ButtonSmall';
-export * from './Logo';
-export * from './Helper';
-export * from './CheckBox';
-export * from './Line';
-export * from './Dot';
-export * from './LineModal';
-export * from './DotModal';
 export * from './Check';
+export * from './CheckBox';
+export * from './Dot';
+export * from './DotModal';
+export * from './Helper';
 export * from './Input';
-export * from './NextButton';
-export * from './BackButton';
+export * from './Line';
+export * from './LineModal';
+export * from './Logo';
 export * from './LoadingIcon';
+export * from './ModalDimmed';
+export * from './ModalIcon';
+export * from './Next';
+export * from './NextButton';
+export * from './SquareButton';
