+++ src/component/uikit/index.ts
@@ -2,3 +2,4 @@ export * from './atom';
 export * from './molecules';
 export * from './organisms';
 export * from './molecules';
+export * from './template';
