+++ src/component/uikit/organisms/index.ts
@@ -1,3 +1,5 @@
 export * from './Header';
 export * from './TabBar';
 export * from './Toast';
+export * from './HelperModal';
+export * from './Modal';
