+++ src/component/uikit/template/LoginCodeTemplate.tsx
@@ -1,8 +1,5 @@
 import styled from 'styled-components';
-import { TabBar } from '@/component/uikit/organisms';
-import { TitleBox } from '@/component/uikit/molecules';
-import { Input } from '@/component/uikit/atom';
-import { ButtonLarge } from '@/component/uikit/atom';
+import { TabBar, TitleBox, Input, ButtonLarge } from '@/component/uikit';
 import { useState, useEffect } from 'react';
 import { TabItem } from '@/types';
 interface Props {
