+++ src/component/uikit/template/index.ts
@@ -1,3 +1,4 @@
+export * from './LoginCodeTemplate';
 export * from './LoginMasterTemplate';
 export * from './PossibleDateTemplate';
 export * from './PossibleTimeTemplate';
